Big time proof of concept on storing frame data outside of store state
Pro: Fixes performance issues that would arrive around frame 3000 before. Not re-creating and diffing a huge array of objects in memory when we really just need to do a read when we change frame.

Con: Code is absolutely awful right now lol, this was more to see if it did fix the problem as predicted. Solutions have to be come up with to keep the replay data all together and ideally to store this somewhere other than the global scope.

Splitting the store into a reactive part (set via setters) and a non-reactive part (directly mutated) seems like a good idea. Interoperability with replay mode is the obvious challenge, but playing in the separate app with limited streaming functionality alllows for some freedom. I think a coherent solution will present itself as the problem is worked through.