Streamline relay process model
Before, there was `StreamsMonitor` to deal with coordinating the stopping of the relay process with the socket process. However, this is not necessary by using OTP primitives. Now, on initialization, `BridgeRelay` processes will link themselves to the appropriate `BridgeSocket` process. This is re-done on restart by making the `BridgeSocket` PID part of the child spec passed to `DynamicSupervisor`. The relay then traps exits to smoothly terminate and notify subscribers when its peer socket exits.

This is resilient to restarts because the re-started relay process will also re-link itself to the socket process, and even if the `terminate/2` callback is not invoked, it will eventually be when the relay is supposed to terminate, and in the meantime it will be restarted by the supervisor.