Proof of concept for displaying viewer + sidebar
Right now this works by dynamically importing the React code when the `bridgeid` attribute is set in the DOM. This is because the SolidJS code is currently written such that it expects a bridge ID to be present on mount, and does not expect it to be changed during runtime. The beginning of this guide was followed: https://stephenbussey.com/2022/04/13/react-in-liveview-how-and-why.html

However, I'm also not sure how it would work if it was designed to react to bridge ID changes. Would it be possible to call a SolidJS setter from app.js and have it behave as expected?

It's worth experimenting with a simple display field rendered in SolidJS, and extrapolate to the overall Viewer component after.