Fix display name length
I noticed when parsing an old netplay replay that the external character IDs for ports 3 and 4 come up as 21. This is very strange, as in hex this represents master hand (26 decimal), which is expected. It's being read in binary so I have no idea what could be happening off the top of my head.