Make Livestream's purpose more explicit
Livestream does not need to do the pubsub forwarding, it just serves to track payload_sizes for the current game while packets are parsed and side-effects are run.

If this wants to be run outside of the BridgeSocket process, this needs to be a separate GenServer instead of a Task because a Task would either necessitate looking up payload_sizes in ETS for every packet, or send payload_sizes back to BridgeSocket asyncronously for that process to track and send back with subsequent packets. This second solution does not feel like it gives ordering guarantees that would be looked for.

The "best" solution by simplicity is just to run the parsing and side-effect logic from the BridgeSocket process via the Streams API, but this feels short-sighted when all the infra for the separate Livestream process already exists. For now, with a small number of streams/viewers, and a small number of side-effects, it would surely work fine, but this process responsibility isolation will proabbly turn useful quickly once many streams are up and more events have side-effects (real-time stats computation).